Add TypeField to its own file instead of TypeLayout file #152
